requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==10.12.0
anthropic==0.34.0
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2
Pillow==10.4.0
